+++ elmyra.ip.access.epo/elmyra/ip/access/epo/static/js/app/views/ops.js
@@ -37,8 +37,6 @@ OpsExchangeDocumentView = Backbone.Marionette.ItemView.extend({
     // actions to run after populating the view
     // e.g. to bind click handlers on individual records
     onDomRefresh: function() {
-        var patent_number = this.model.attributes.get_patent_number();
-        opsChooserApp.basketView.link_document(patent_number);
     },
 
 });